+++ src/components/Follows/Follows.tsx
@@ -5,7 +5,7 @@ import { CenteredLoader } from "@/components/CenteredLoader/CenteredLoader";
 import { UsersList } from "@/components/UsersList/UsersList";
 import type { User } from "@/types/user";
 import { baseURL } from "@/utils/baseUrl";
-import { fetcher } from "@/utils/fetcher";
+import { fetcher } from "@/utils/api";
 
 interface FollowsProps {
 	currentUserId: string;
